server/requirements.txt
cryptography==41.0.0
face_recognition==1.3.0
Flask==2.3.2
Flask_Bcrypt==1.0.1
Flask_Cors==3.0.10
Flask_JWT_Extended==4.5.2
Flask_Mail==0.9.1
flask_simple_crypt==0.3.3
numpy==1.23.1
opencv_python==4.6.0.66
Pillow==9.5.0
pymongo==4.3.3
pyotp==2.8.0
python-dotenv==1.0.0